NixOS layout: add manual, remove options+packages
The previous layout lacked a manual link in the navigation bar at
the top, which suggested by analogy with Nix/Nixpkgs/NixOps that
NixOS did not in fact have a manual. To make space for it, I've
removed the "Packages" and "Options" links, which I've made sure
to still mention on the Support page along with the manual.