Splits the navigation in two navbars...
 * Makes the different projects more obvious.
 * Marks a clearer hierarchy in the website.

In addition, since the DOM had to be manipulated, and we
have a bit more space:

 * Re-adds github and twitter links for nixos project.

The dropdown for the projects is, as a side-effect from manipulating
the DOM, now gone from the mobile view. It is now a flat list of
projects, followed by the page list of that specific section.