Fix RSS encoding of news items
The current news-rss.xml fails http://www.feedvalidator.org for a couple
of reasons, making feed readers like Liferea misparse it. This patch
fixes one of those issues.

As @edolstra noted in #90, HTML is not allowed directly in the RSS
`<description>` tag. Instead it's supposed to be a text node, where the
text can be parsed as HTML. So e.g. an `<a>` tag needs to be written in
the feed as `&lt;a&gt;`.

This could be handled more easily in an Atom feed, which allows tags
from namespaces like XHTML to be embedded directly as long as the
namespaces are declared properly.

This could also be simpler if xsltproc supported the XPath Functions 3.0
standard, which defines a `serialize` function that does what the
mode="serialize" template does in this patch.