Fix pubDate timestamps in news RSS feed
The current news-rss.xml fails http://www.feedvalidator.org for a couple
of reasons, making feed readers like Liferea misparse it. This patch
fixes one of those issues.

The RSS pubDate element is required to be an RFC-822 date-time. The
entries in news.xml did not conform to that specification for two
reasons: a missing comma, and swapping the month and day fields.

So I fixed both of those issues in news.xml, and then fixed news.xsl to
extract the date substring at its new offset.

This does mean that the HTML version of the news has its month and day
swapped now, and frankly I liked it better in the previous month-day
order.

If desired the previous output can be recovered, by either continuing
using the substring-pasting approach but chopping up the substrings
further, or changing news.xml to use ISO 8601 date-times and using the
http://exslt.org/date/ extension functions.