Fix news RSS item properties
The item's <link> tag should point to an HTML version of the same
content as the item. However, there is no URL for each item in the
current site, so there's no good place to link to. Fortunately, the link
tag is not required when a description is present, so we can just drop
that tag entirely.

Also, each item should have some sort of unique ID in the <guid> tag.
If a feed reader sees an item that's different from any item it has seen
before, then the guid allows the reader to distinguish between new items
and items that have been edited.

http://www.feedvalidator.org/docs/warning/MissingGuid.html

Since it has almost never happened that two items were posted with the
same pubDate, I've chosen to use the pubDate itself as the guid. There
were three duplicates, but I've made them unique by adding one second to
the pubDates of the three duplicates that appear earlier in the file, so
that sorting by descending pubDate would leave the file order unchanged.