Mark news RSS feed as "complete" per RFC5005
First, I removed the maxItem limit from the RSS feed. (It's still used
for the HTML version.) The maxItem limit of 1000 was effectively
infinite at the rate news items are being published, and each item is
quite small so there isn't much bandwidth cost to publishing a lot of
them anyway.

Second, since the feed does contain every entry ever posted, it is a
valid "complete feed" in the sense of RFC5005 section 2, so add the tag
marking it as such.

If the RSS feed ever gets too large, then RFC5005 section 4 describes
how to paginate it into archived feeds.