Updates landing page to be more project agnostic
From the nixos.org main page, there is was no way to quickly glance at
the moving parts of the NixOS project, that is Nix and Nixpkgs. Only
NixOS was represented.

This is a step into fixing an issue that @garbas brought forward at the
last NixCon, where the landing page of NixOS.org and the nix landing
page lacks in "actionability".

From the new nixos.org main page, it is easier to go straight to the Nix
specific section of the site. In a follow up work, the main Nix web page
can be re-worked to "show off" more what Nix is about.