Decouple build nixpkgs from released nixpkgs
The eager update to use 20.03 to *build* the flake accidentally made the
site use the 20.03 manual for a while.

This change allows us to move forward the infrastructure of the site
in preparation for a release, without actually using that version's
manual. This accidental use of an unreleased release should not happen
anymore.

Additionally, it's not much of an issue, the list of AMIs could have
been using the wrong list (of an unreleased branch), though as it
requires manual involvement of the team it wouldn't have been an issue.

Fixes #400.