site-styles: Prevent ridiculously small device issues
Yes, we're scaling down the site, but we are **not** making it worse.
This is because the site structure will not be glitched out by the weird
proportions of those devices, and the users can still zoom-in.

In fact, the actual pixel density of their displays allows the user to
simply bring the phone closer to their face also!