common.tt: Add facility to embed a demo preview SVG
This *strongly assumes* that the SVG will be available, and also assumes
that the DOM and styles are ready to embed the SVG at this location.

Additionally, yet unwritten, another component will be needed at the
location where the "pane" with the actual demo will be shown.

Finally, JavaScript will be used to make the pane show up properly when
activated.