copy-nix-dev-tutorials: use printf to fix title
The title was using what amounted to:

   echo '[% $title %]' > file.tt

Which put the literal `$title` perl variable, which in Template Toolkit,
an undefined variable is equivalent to an empty string.

This made the <title /> of the page be "NixOS - Guides - ".

This fixes that by using the better `printf` building for building
strings to print.

Additionally we can drop the empty echoes and use `\n` as printf honours
them.