site-styles: review margin distribution in footer
The previous way it was distributed made undesirable spacing when the
display was narrow.

Instead of using the natural margins of those elements, we instead set a
specific margin at the top of the upper part. This means that when the
elements are placed in a vertical manner there is less spacing between
element; the desired goal.