news.xml: fix links as far as possible
In order to honor the history of Nix(OS), this is an effort to restore
the announcement page to former glory. It is really quite fascinating,
how basically every link before 2010 is broken by now.

Some links are unfortunately beyond rescue, including University Delft
websites that were never archived and now point to a Github 404 …

Things left to do:
- Change all pdfs to archive links. Eelco’s papers still work, but
  they shouldn’t change and the links will break eventually. Better
  safe than sorry.
- Fix the release pages. From around 2010, a grave sin was commited
  and most links point directly to build outputs of hydra.nixos.org,
  which are obviously all broken by now. I found that most should be
  easy to recover by pointing them to `releases.nixos.org`. Others
  might be lost.
- Think about how to link to historical release documents that use
  relative links right now. The sections might be restructured in the
  future, breaking the anchors.
- Archive videos of talks we point to. At least one is on Youtube.

Most external historical links should be back to “working”, for
various definitions of working.