Label buttons to the manuals with the manual names
This makes it clearer what the 3 'Read more...' buttons actually point to, and
that they are pointing to 3 different manuals.

Seems to fit fine on all zoom levels / window sizes I tried, and is helpful
especially in the mobile view where the 3 sections are actually below each
other rather than next to each other.