site-styles: Fix text-replacement with media queries
It's simply impossible to use `:extend` with media queries. This is
because the rules are defined for a different media query!

As LESS *only* appends selectors to the other one, it doesn't since the
media queries don't match.

Sure, this means a little bit of repetition in the styles, but eh, what
can we do?