collapse: prefer using implicit `.with-js` class
By prefixing `&` we can make styles in a namespace apply according to
external selectors.

Here `html.with-js &` makes is so that when the selector is expanded,
`&` is replaced with the parent selector (`.collapse`), so it ends up
being `html.with-js .collapse`.

This does have one caveat. The previous implementation would keep the
"un-enabled" variant for any collapse not already converted if the code
ended up stopping due to a crash. This is not something that we should
concern ourselves with, as if it crashes there's more to worry about.