js: move page-specific code
Ideally we'll prefer implementing a generalized behaviour instead of
page-specific JS. But when it's needed, put it at the end so the file is
easier to go through.

Maybe we should invest some time into figuring out a palatable "packer"
for the JavaScript bits? Probably to be looked at later when we choose
the new tooling for the website.