js: Prefer attributes accessors to build HTML
Using `$("<a />")` is perfectly fine. Even `$("<a class='something' />").
But the moment we want to include variable content, prefer using `.attr`
rather than hoping the content will not destroy our string's HTML.

Sure, we control the DOM, and *it shouldn't happen*. But as we're making
this generic enough, we can make it right at the same time.

Furthermore, this, in my opinion, removes the cognitive load of reading
the escaped and concatenated string. (This is why the double quotes were
replaced with single quotes).