js: collapse $titles -> $articles
This, coupled with the next change,  reduces confusion when mentally
mapping the model of the collapse.

We don't *really* care about titles, we care about the articles.

This is basically a useless diversion at this point in time, but when
we need to refer to the article in the future, we're already using the
article instead of targeting deeply for its title.