js: Fixes countdown logic
First, we should ingest the date as UTC.

Then, we should definitely not use "new Date()" on the result of a
subtraction as a way to do date maths. We could be crossing into another
timezone due to daylight saving and such!

This was tested by comparing the actual time left to the event with what
is on the page.