site-styles: Fix asciinema breaking width by removing aspect ratio code
The reason there was that aspect ratio thing there was to, if I remember
right, make the whole player fit in the height of a side-ways mobile
device.

Though I think I must have forgotten to finish implementing this, as
instead it just resolved to using the default font sizes.

Rather, let's do what works fine for every other sizes: a ratio of the
device width.