NixOS Download: recommend GNOME
The GNOME ISO correctly detects and configures HiDPI scaling.
More, it supports changing the scale. The Plasma Desktop ISO
doesn't detect and configure the scale automatically, and doesn't
allow you to change it without restarting. This makes the GNOME
ISO significantly more accessible for users on laptops with HiDPI
displays.