explore page: Rephrase third bullet
The second paragraph was hard to understand. How do you fix "it" (a
broken configuration run) locally before it happens? My interpretation
is it's about fixing problems that would cause CI to fail by testing
locally. And the nix benefit would be that the local and CI environments
will agree on the versions of testing and other tools which would run.