Rework build for CSS
This is done to remove the SVG embed step; the LESS files will use the
new `read-file` plugin to embed the SVG during build.

This is not "correct", though, as this is currently not optimizing the
SVG files.

Optimizing SVGs *for final build* will be **required** for merging this.
Optimizing for development should be done, but requires some thinking to
work "around" the fact this is built using a Makefile.