stepping back from documentation team
Thanks for all the work! I see myself already be saturated with writing
documentation for my own project. I fail to keep track with all the work
that the documentation team and the community is doing and hence I feel
like I do not have much to contribute to the regular meetings.

In case someone wonders, this is not related to
https://discourse.nixos.org/t/parting-from-the-documentation-team/24900
in any way. I have not followed what happend there and this decission
has been already made before.