Display Nix versioned manuals
Manuals for all the version in the nix_versions.json files are
generated, but only the 3 previous are displayed on the website.

Includes a script, update-nix-versions.py, to update nix_versions.json.
New versions are automatically added. The script also updates the patch
version (e.g.: if latest 2.13 version is 2.13.3, the script will replace
that version with 2.13.4)