remove inclusion of nix.dev articles
it's too brittle and breaks on every little change.
we don't have the infrastructure yet to set up deal with redirects.

we may want to go away from rolling releases of documentation to avoid
that type of problem in the future, and naturally stay in sync with
Nixpkgs and NixOS releases.