workflows/cron: Pin Nix to 2.19.2
The current default for cachix/install-nix-action@v24 is 2.19.1, which
annoyingly completely removed support for `--update-input`, which the
homepages udpate script was relying upon.
Only 2.19.2 reinstated support for that flag with a deprecation notice.

Also by pinning we prevent such problems in the future.
Notably in this case this caused the docs on nixos.org to not update
**for 2 months**.