Astro/catchup site content since diverge (#1211)
See meeting notes (31 Jan 2024). Captures many of the updates to the
site since `main` was forked from `master` to start the site refactor.
Commit messages capture the PR that was migrated.

From #1182. There isn't a Tokodon icon in simple icons so the Mastodon
icon was used. This creates 2 Mastodon icons on the community page.

From #1152. Since the teams page had a layout change. I organized the
content as best as possible.

From #1159. Since the teams page had a layout change. I organized the
content as best as possible.