Set asciinema fit to 'width'
https://docs.asciinema.org/manual/player/options/#fit

It seems like we need to make a tradeoff between fitting all the text in
or having a larger font. I'm betting on showing all the content being
the more important of the two and that we can have terminal replays with fewer columns
if we really want larger font. And it's easier to rely on asciinema's
player to be responsive than to adjust font size manually ourselves.

Also deleted `terminalFontSize` per docs that this is only relevant if
fit is set to false.

Temporarily commented out `terminalFontFamily` as this is causing some
sizing issues too