Use font loading API to block on font loading before mounting AsciinemaPlayer
Even with `fit: 'width'`, trying to use Fira Code Variable was causing
asciinema player elements to cut text off (at least in Firefox and Safari).

Per the note on the bottom of https://docs.asciinema.org/manual/player/fonts/
```
The player performs measurement of font metrics (width/height) when it mounts in the page, therefore it's highly recommended to ensure chosen font is already loaded before calling create. This can be achieved by using CSS Font Loading API as seen in the above example.
```

I believe this was the cause of the issue because the AsciinemaPlayer
can't compute the width correctly until the font is loaded.

To test this, it's best to observe the network tab for font transfes with
Disable Caches checked to verify that the layout issue happens when the
font isn't retrieved from the cache.