fix: rework platform introductions
Personally, I think this is what the community page should say.
The wording could be modified, but I see no good reason to
inlcude a big list of unofficial fora that may or may not reflect
well on Nix as a whole. In terms of PR / marketing, it seems a
basic mistake to point people elsewhere.