Fix command in nixpkgs docker instructions (#1664)
Closes: #1415 

I tested the commands in the docs on my local and they failed then I
tested the edited commands in the issue reply by
[@mcdonc](https://github.com/mcdonc) and they worked.
The original fails because we mount `/nixpkgs/` to `/nixpkgs` in the
previous command `docker run -it -v $(pwd)/nixpkgs:/nixpkgs nixos/nix`
instead of '/'. I think this is intentional, however it means that
`default.nix` is located at `/nixpkgs/default.nix` and has to be
specified in the `nix-build` command.