add unstable iso image download (#1677)
This MR adds the download links to the unstable image again.

I chose to add an unstable ISO link to the NixOS sidebar on the download
page (see screenshot below).

![Screenshot From 2025-03-03
17-35-00](https://github.com/user-attachments/assets/d80ffbf8-c7e4-4529-9f7d-ddfa1da2921d)

If this is confusing, we could also add a separate section for the
unstable downloads. I am not 100% sure what the best way to approach
this is.

Fixes #635.

---------

Co-authored-by: Thilo Billerbeck <thilo.billerbeck@officerent.de>