Revert "add unstable iso image download" (#1701)
Reverts NixOS/nixos-homepage#1677

See: https://github.com/NixOS/nixpkgs/pull/355893

Unstable will be changing the layout of the ISOs somewhat in the next
channel update. Importantly, there will no longer be separate GNOME and
Plasma ISOs. Instead there will be a single combined ISO where either
can be selected in the boot menu. This makes the incoming unstable
update incompatible with the recently merged PR that adds unstable ISOs
to the homepage.

The simplest course of action seems to be to revert the addition to the
homepage and reintegrate it when the homepage updates for 25.05, when
these changes will be live for both stable and unstable.

@flyfloh My sincere apologies. I should have communicated the imminent
changes to the unstable ISO before your PR was merged. I will help
reintegrate it once 25.05 comes around.