Restructure project files (#1714)
This PR helps resolve some build time issues and live reload hickups by
moving the core part of the website into its own workspace, called
`/core`. Thus, the website's source does not conflict with anything that
is introduced by the flake and its external builds.

In short
- move the core website into `/core`
- Introduce npm workspaces
- more clearly separate the core website build and external builds such
as the manual