Fix double clicks on banner buttons (#1843)
When clicking on an `<a>` element in a banner we open the link twice,
once from the <a> and once from JavaScript. This is easily visible
currently with the big "Take the Survey" button for the Nix Community
Survey banner. I don't know if this solution is the best approach, it
seems logical to me but I haven't used astro before.