Add devices page (#1836)
The hardware team is looking to add a page which lists devices that are
supported. The support falls under 3 "tiers":
- Ships with NixOS
- Official Support
- Community Support

Ships with NixOS is a bunch of devices which ship with NixOS out of the
box. This means the user has to only configure the system to their
liking. No need for any installation ISO or changing secure boot
options.

Official support are devices which the hardware team is able to
"guarantee" support. These devices have been received by the hardware
team and checked to be supported.

Community supported devices are ones which the hardware team is not able
to provide support. A good example is Apple Silicon machines. Those
machines are handled by a NixOS community repo.

The goal here is to incentivize vendors to ship with NixOS, encourage
sponsorships from vendors, and provide users a central way to find out
about hardware support. This has already been shown to @djacu and @Mic92
at NixCon.

---------

Co-authored-by: Thilo Billerbeck <thilo.billerbeck@officerent.de>