research: add master-thesis about finding non-reproducible derivations in a derivations build closure (#1926)
I used my date of submission to order it in compared with others date
published, please let me know if that should rather be the actual date
of self-publication/uploading to zenodo.

In the document I did not use an abstract so I did not add one into this
PR. I see this potentially being impractical but I am also not sure
about adding something in there that is not actually in the document.

@mschwaig